feat: tighten opportunitiesSlice types and move them into their own module
