+++ src/state/selectors.ts
@@ -4,7 +4,7 @@ import type { AccountId, AssetId, ChainId } from '@shapeshiftoss/caip'
 import createCachedSelector from 're-reselect'
 
 import type { ReduxState } from './reducer'
-import type { LpId, StakingId, UserStakingId } from './slices/opportunitiesSlice/opportunitiesSlice'
+import type { LpId, StakingId, UserStakingId } from './slices/opportunitiesSlice/types'
 import type { PubKey } from './slices/validatorDataSlice/validatorDataSlice'
 
 // List of all the params filter consumed with selectParamFromFilter
