+++ src/state/slices/opportunitiesSlice/opportunitiesSlice.ts
@@ -1,80 +1,22 @@
 import { createSlice } from '@reduxjs/toolkit'
 import { createApi } from '@reduxjs/toolkit/query/react'
-import type { AccountId, AssetId } from '@shapeshiftoss/caip'
-import type { DefiType } from 'features/defi/contexts/DefiManagerProvider/DefiCommon'
 import { DefiProvider } from 'features/defi/contexts/DefiManagerProvider/DefiCommon'
 import merge from 'lodash/merge'
 import { BASE_RTK_CREATE_API_CONFIG } from 'state/apis/const'
-import type { Nominal } from 'types/common'
 
 import {
   getMetadataResolversByDefiProviderAndDefiType,
   getUserDataResolversByDefiProviderAndDefiType,
 } from './resolvers/utils'
-
-export type OpportunityMetadata = {
-  apy: string
-  assetId: AssetId
-  provider: DefiProvider
-  tvl: string
-  type: DefiType
-  underlyingAssetIds: readonly [AssetId, AssetId]
-}
-
-// User-specific values for this opportunity
-export type UserStakingOpportunity = {
-  // The amount of farmed LP tokens
-  stakedAmountCryptoPrecision: string
-  // The amount of rewards available to claim for the farmed LP position
-  rewardsAmountCryptoPrecision: string
-}
-
-// The AccountId of the staking contract in the form of chainId:accountAddress
-export type StakingId = Nominal<string, 'StakingId'>
-// The AccountId of the LP contract in the form of chainId:accountAddress
-export type LpId = Nominal<string, 'LpId'>
-// The unique identifier of an lp opportunity in the form of UserAccountId*StakingId
-export type UserStakingId = `${AccountId}*${StakingId}`
-
-export type OpportunitiesState = {
-  lp: {
-    byAccountId: Record<AccountId, LpId[]> // a 1:n foreign key of which user AccountIds hold this LpId  [id: AccountId]: LpId[]
-    byId: Record<LpId, OpportunityMetadata>
-    ids: LpId[]
-  }
-  // Staking is the odd one here - it isn't a portfolio holding, but rather a synthetic value living on a smart contract
-  // Which means we can't just derive the data from portfolio, marketData and other slices but have to actually store the amount in the slice
-  userStaking: {
-    byId: Record<UserStakingId, UserStakingOpportunity>
-    ids: UserStakingId[]
-  }
-
-  staking: {
-    // a 1:n foreign key of which user AccountIds hold this StakingId
-    byAccountId: Record<AccountId, StakingId[]>
-    byId: Record<StakingId, OpportunityMetadata>
-    ids: StakingId[]
-  }
-}
-
-export type OpportunityMetadataById = OpportunitiesState['lp' | 'staking']['byId']
-export type OpportunityDataById = OpportunitiesState['lp' | 'staking']['byAccountId']
-
-export type GetOpportunityMetadataInput = {
-  opportunityId: LpId | StakingId
-  opportunityType: 'lp' | 'staking'
-  defiType: DefiType
-}
-
-export type GetOpportunityUserDataInput = {
-  accountId: AccountId
-  opportunityId: LpId | StakingId
-  opportunityType: 'lp' | 'staking'
-  defiType: DefiType
-}
-
-type GetOpportunityMetadataOutput = any
-type GetOpportunityUserDataOutput = any
+import type {
+  GetOpportunityMetadataInput,
+  GetOpportunityMetadataOutput,
+  GetOpportunityUserDataInput,
+  GetOpportunityUserDataOutput,
+  OpportunitiesState,
+  OpportunityDataById,
+  UserStakingId,
+} from './types'
 
 export const initialState: OpportunitiesState = {
   lp: {
@@ -100,18 +42,18 @@ export const opportunities = createSlice({
     clear: () => initialState,
     upsertOpportunityMetadata: (
       draftState,
-      { payload }: { payload: { metadata: OpportunityMetadataById; type: 'lp' | 'staking' } },
+      { payload }: { payload: GetOpportunityMetadataOutput },
     ) => {
       draftState[payload.type].byId = {
         ...draftState[payload.type].byId,
-        ...payload.metadata,
+        ...payload.byId,
       }
-      draftState[payload.type].ids = Array.from(new Set([...Object.keys(payload.metadata)]))
+      draftState[payload.type].ids = Array.from(new Set([...Object.keys(payload.byId)]))
     },
     // TODO: testme
     upsertOpportunityAccounts: (
       draftState,
-      { payload }: { payload: { byAccountId: OpportunityDataById; type: 'lp' | 'staking' } },
+      { payload }: { payload: GetOpportunityUserDataOutput },
     ) => {
       draftState[payload.type].byAccountId = merge(
         draftState[payload.type].byAccountId,
